+++ src/Type/Definition/HasFieldsTypeImplementation.php
@@ -8,7 +8,10 @@ use GraphQL\Utils\Utils;
 
 use function array_keys;
 
-abstract class TypeWithFields extends Type implements HasFieldsType
+/**
+ * @see HasFieldsType
+ */
+trait HasFieldsTypeImplementation
 {
     /**
      * Lazily initialized.