+++ src/Type/Definition/ListOfType.php
@@ -6,8 +6,6 @@ namespace GraphQL\Type\Definition;
 
 use GraphQL\Type\Schema;
 
-use function is_callable;
-
 /**
  * @template OfType of Type
  */
@@ -17,7 +15,7 @@ class ListOfType extends Type implements WrappingType, OutputType, NullableType,
      * @var Type|callable
      * @phpstan-var OfType|callable(): OfType
      */
-    public $ofType;
+    private $wrappedType;
 
     /**
      * @param Type|callable $type
@@ -25,30 +23,30 @@ class ListOfType extends Type implements WrappingType, OutputType, NullableType,
      */
     public function __construct($type)
     {
-        $this->ofType = is_callable($type)
-            ? $type
-            : Type::assertType($type);
+        $this->wrappedType = $type;
     }
 
     public function toString(): string
     {
-        return '[' . $this->getOfType()->toString() . ']';
+        return '[' . $this->getWrappedType()->toString() . ']';
     }
 
     /**
      * @phpstan-return OfType
      */
-    public function getOfType(): Type
+    public function getWrappedType(): Type
     {
-        return Schema::resolveType($this->ofType);
+        return Schema::resolveType($this->wrappedType);
     }
 
-    public function getWrappedType(bool $recurse = false): Type
+    public function getInnermostType(): NamedType
     {
-        $type = $this->getOfType();
+        $type = $this->getWrappedType();
+        while ($type instanceof WrappingType) {
+            $type = $type->getWrappedType();
+        }
 
-        return $recurse && $type instanceof WrappingType
-            ? $type->getWrappedType($recurse)
-            : $type;
+        /** @var Type&NamedType $type known because we unwrapped all the way down */
+        return $type;
     }
 }