+++ src/Type/Definition/NamedType.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace GraphQL\Type\Definition;
 
+use GraphQL\Error\Error;
 use GraphQL\Language\AST\Node;
 use GraphQL\Language\AST\TypeDefinitionNode;
 use GraphQL\Language\AST\TypeExtensionNode;
@@ -17,9 +18,20 @@ use GraphQL\Language\AST\TypeExtensionNode;
  * | EnumType
  * | InputObjectType;
  *
+ * @property string $name
+ * @property string|null $description
  * @property (Node&TypeDefinitionNode)|null $astNode
  * @property array<int, Node&TypeExtensionNode> $extensionASTNodes
  */
 interface NamedType
 {
+    /**
+     * @throws Error
+     */
+    public function assertValid(): void;
+
+    /**
+     * Is this type a built-in type?
+     */
+    public function isBuiltInType(): bool;
 }