+++ src/Type/Definition/NonNull.php
@@ -6,19 +6,24 @@ namespace GraphQL\Type\Definition;
 
 use GraphQL\Type\Schema;
 
+/**
+ * @phpstan-type WrappedType (NullableType&Type)|callable():(NullableType&Type)
+ */
 class NonNull extends Type implements WrappingType, OutputType, InputType
 {
-    /** @var (NullableType&Type)|callable():(NullableType&Type) */
-    private $ofType;
+    /**
+     * @var Type|callable
+     * @phpstan-var WrappedType
+     */
+    private $wrappedType;
 
     /**
-     * code sniffer doesn't understand this syntax. Pr with a fix here: waiting on https://github.com/squizlabs/PHP_CodeSniffer/pull/2919
-     * phpcs:disable Squiz.Commenting.FunctionComment.SpacingAfterParamType
-     * @param (NullableType&Type)|callable():(NullableType&Type) $type
+     * @param Type|callable $type
+     * @phpstan-param WrappedType $type
      */
     public function __construct($type)
     {
-        $this->ofType = $type;
+        $this->wrappedType = $type;
     }
 
     public function toString(): string
@@ -29,22 +34,20 @@ class NonNull extends Type implements WrappingType, OutputType, InputType
     /**
      * @return NullableType&Type
      */
-    public function getOfType(): Type
+    public function getWrappedType(): Type
     {
         // @phpstan-ignore-next-line generics in Schema::resolveType() are not recognized correctly
-        return Schema::resolveType($this->ofType);
+        return Schema::resolveType($this->wrappedType);
     }
 
-    /**
-     * @return NullableType&Type
-     */
-    public function getWrappedType(bool $recurse = false): Type
+    public function getInnermostType(): NamedType
     {
-        $type = $this->getOfType();
+        $type = $this->getWrappedType();
+        while ($type instanceof WrappingType) {
+            $type = $type->getWrappedType();
+        }
 
-        // @phpstan-ignore-next-line might return another type intermediary, but will always match the return type when done with recursion
-        return $recurse && $type instanceof WrappingType
-            ? $type->getWrappedType($recurse)
-            : $type;
+        /** @var Type&NamedType $type known because we unwrapped all the way down */
+        return $type;
     }
 }