+++ tests/Type/DefinitionTest.php
@@ -226,9 +226,11 @@ class DefinitionTest extends TestCase
 
         self::assertEquals($this->blogSubscription, $schema->getSubscriptionType());
 
-        $sub = $this->blogSubscription->getField('articleSubscribe');
-        self::assertEquals($sub->getType(), $this->blogArticle);
-        self::assertEquals($sub->getType()->name, 'Article');
+        $sub     = $this->blogSubscription->getField('articleSubscribe');
+        $subType = $sub->getType();
+        self::assertInstanceOf(ObjectType::class, $subType);
+        self::assertEquals($subType, $this->blogArticle);
+        self::assertEquals($subType->name, 'Article');
         self::assertEquals($sub->name, 'articleSubscribe');
     }
 
@@ -603,11 +605,11 @@ class DefinitionTest extends TestCase
 
         $blogFieldReturnType = $user->getField('blogs')->getType();
         self::assertInstanceOf(NonNull::class, $blogFieldReturnType);
-        self::assertSame($blog, $blogFieldReturnType->getWrappedType(true));
+        self::assertSame($blog, $blogFieldReturnType->getInnermostType());
 
         $ownerFieldReturnType = $blog->getField('owner')->getType();
         self::assertInstanceOf(NonNull::class, $ownerFieldReturnType);
-        self::assertSame($user, $ownerFieldReturnType->getWrappedType(true));
+        self::assertSame($user, $ownerFieldReturnType->getInnermostType());
     }
 
     public function testInputObjectTypeAllowsRecursiveDefinitions(): void
@@ -798,11 +800,10 @@ class DefinitionTest extends TestCase
             'name'   => 'SomeObject',
             'fields' => [['field' => Type::string()]],
         ]);
-        $this->expectException(InvariantViolation::class);
-        $this->expectExceptionMessage(
-            'SomeObject fields must be an associative array with field names as keys or a ' .
-            'function which returns such an array.'
-        );
+
+        $this->expectExceptionObject(new InvariantViolation(
+            'SomeObject fields must be an associative array with field names as keys or a function which returns such an array.'
+        ));
         $objType->getFields();
     }
 
@@ -817,11 +818,10 @@ class DefinitionTest extends TestCase
                 return [['field' => Type::string()]];
             },
         ]);
-        $this->expectException(InvariantViolation::class);
-        $this->expectExceptionMessage(
-            'SomeObject fields must be an associative array with field names as keys or a ' .
-            'function which returns such an array.'
-        );
+
+        $this->expectExceptionObject(new InvariantViolation(
+            'SomeObject fields must be an associative array with field names as keys or a function which returns such an array.'
+        ));
         $objType->getFields();
     }
 
@@ -846,30 +846,6 @@ class DefinitionTest extends TestCase
         ]);
         // Should not throw:
         $objType->assertValid();
-    }
-
-    // rejects an Object type with incorrectly typed field args
-
-    /**
-     * @see it('does not allow isDeprecated without deprecationReason on field')
-     */
-    public function testDoesNotAllowIsDeprecatedWithoutDeprecationReasonOnField(): void
-    {
-        $OldObject = new ObjectType([
-            'name'   => 'OldObject',
-            'fields' => [
-                'field' => [
-                    'type'         => Type::string(),
-                    'isDeprecated' => true,
-                ],
-            ],
-        ]);
-
-        $this->expectException(InvariantViolation::class);
-        $this->expectExceptionMessage(
-            'OldObject.field should provide "deprecationReason" instead of "isDeprecated".'
-        );
-        $OldObject->assertValid();
     }
 
     // Object interfaces must be array
@@ -1144,10 +1120,10 @@ class DefinitionTest extends TestCase
      */
     public function testRejectsAnInterfaceTypeWithAnIncorrectTypeForResolveType(): void
     {
-        $this->expectException(InvariantViolation::class);
-        $this->expectExceptionMessage(
-            'AnotherInterface must provide "resolveType" as a function, but got: instance of stdClass'
-        );
+        // Slightly deviating from the reference implementation in order to be idiomatic for PHP
+        $this->expectExceptionObject(new InvariantViolation(
+            'AnotherInterface must provide "resolveType" as a callable, but got: instance of stdClass'
+        ));
 
         $type = new InterfaceType([
             'name'        => 'AnotherInterface',
@@ -1209,10 +1185,10 @@ class DefinitionTest extends TestCase
      */
     public function testRejectsAnUnionTypeWithAnIncorrectTypeForResolveType(): void
     {
-        $this->expectException(InvariantViolation::class);
-        $this->expectExceptionMessage(
-            'SomeUnion must provide "resolveType" as a function, but got: instance of stdClass'
-        );
+        // Slightly deviating from the reference implementation in order to be idiomatic for PHP
+        $this->expectExceptionObject(new InvariantViolation(
+            'SomeUnion must provide "resolveType" as a callable, but got: instance of stdClass'
+        ));
         $this->schemaWithFieldType(
             new UnionType([
                 'name'        => 'SomeUnion',
@@ -1378,10 +1354,10 @@ class DefinitionTest extends TestCase
      */
     public function testRejectsAnObjectTypeWithAnIncorrectTypeForIsTypeOf(): void
     {
-        $this->expectException(InvariantViolation::class);
-        $this->expectExceptionMessage(
-            'AnotherObject must provide "isTypeOf" as a function, but got: instance of stdClass'
-        );
+        // Slightly deviating from the reference implementation in order to be idiomatic for PHP
+        $this->expectExceptionObject(new InvariantViolation(
+            'AnotherObject must provide "isTypeOf" as a callable, but got: instance of stdClass'
+        ));
         $this->schemaWithFieldType(
             new ObjectType([
                 'name'     => 'AnotherObject',